Refactor article controller and add articles helper
