+++ app/helpers/articles_helper.rb
@@ -1,0 +1,2 @@
+module ArticlesHelper
+end
